moved PlaceDetail from components to screens folder; updated FindPlace screen with functionality of navigation to the new PlaceDetail screen
